sprint 5 issues fixed  and call configuration apis

+++ ViewControllers/MyFavourites/MyFavouritesViewController.swift
@@ -11,6 +11,8 @@ import GooglePlaces
 
 class MyFavouritesViewController: UIViewController {
 
+    @IBOutlet weak var buttonPrayerPlaces: UIButton!
+    @IBOutlet weak var buttonRestaurant: UIButton!
     @IBOutlet weak var imageViewNoAddressFound: UIImageView!
     @IBOutlet weak var buttonAddNewAddress: UIButton!
     @IBOutlet weak var viewTitle: UIView!
@@ -20,7 +22,7 @@ class MyFavouritesViewController: UIViewController {
     @IBOutlet weak var viewNoDataFoundBackGround: UIView!
     var favouritePageNumber = 1
 
-    var selectedAddressIndex: Int? = 0 {
+    var selectedIndex: Int? = 0 {
         didSet {
             tableView.reloadData()
         }
@@ -33,24 +35,23 @@ class MyFavouritesViewController: UIViewController {
             
             if tableView.visibleCells.count == 0 {
                 viewNoDataFoundBackGround.isHidden = false
+                tableView.isHidden = true
             }
             else {
                 viewNoDataFoundBackGround.isHidden = true
+                tableView.isHidden = false
             }
         }
     }
     
-    var modelDeleteUserAddressResponse: ModelDeleteUserAddressResponse? {
+    var modelPostFavouriteDeleteResponse: FindHalalFoodCell.ModelPostFavouriteDeleteResponse? {
         didSet {
-            if modelDeleteUserAddressResponse?.success ?? false {
-                getFavouriteByUser()
-//                showAlertCustomPopup(title: "Success", message: modelDeleteUserAddressResponse?.message ?? "", iconName: .iconSuccess) { _ in
-//                    self.tableView.reloadData()
-//                    self.getFavouriteByUser()
-//                }
+            print(modelPostFavouriteDeleteResponse as Any)
+            if modelPostFavouriteDeleteResponse?.success ?? false {
+                tableView.reloadData()
             }
             else {
-                showAlertCustomPopup(title: "Error", message: modelDeleteUserAddressResponse?.message ?? "", iconName: .iconError)
+                self.showAlertCustomPopup(title: "Error!", message: modelPostFavouriteDeleteResponse?.message ?? "", iconName: .iconError)
             }
         }
     }
@@ -62,6 +63,24 @@ class MyFavouritesViewController: UIViewController {
         viewTitle.radius(radius: 12)
         getFavouriteByUser()
         imageViewNoAddressFound.isHidden = false
+        viewBottomLinePrayerPlaces.isHidden = true
+        buttonRestaurant.tag = 1
+    }
+
+    @IBOutlet weak var viewBottomLineRestaurants: UIView!
+    @IBOutlet weak var viewBottomLinePrayerPlaces: UIView!
+    @IBAction func buttonRestaurant(_ sender: Any) {
+        viewBottomLinePrayerPlaces.isHidden = true
+        viewBottomLineRestaurants.isHidden = false
+        buttonRestaurant.tag = 1
+        tableView.reloadData()
+    }
+    
+    @IBAction func buttonPrayerPlaces(_ sender: Any) {
+        viewBottomLinePrayerPlaces.isHidden = false
+        viewBottomLineRestaurants.isHidden = true
+        buttonRestaurant.tag = 0
+        tableView.reloadData()
     }
     
     @IBAction func buttonBack(_ sender: Any) {
@@ -93,16 +112,27 @@ class MyFavouritesViewController: UIViewController {
     }
     
     func deleteUserAddress(index: Int) {
-//        let parameters: Parameters = [
-//            "id": modelGetUserAddressResponse?.userAddressesResponseData?[index].id ?? ""
-//        ]
-//        let id = modelGetUserAddressResponse?.userAddressesResponseData?[index].id ?? ""
-//        
-//        let url = "\(APIsName.name.deleteuseraddress.rawValue)/\(id)"
-//        APIs.deleteAPI(apiName: url, parameters: nil, methodType: .delete, viewController: self) { responseData, success, errorMsg in
-//            let model: ModelDeleteUserAddressResponse? = APIs.decodeDataToObject(data: responseData)
-//            self.modelDeleteUserAddressResponse = model
-//        }
+        var deleteID = ""
+        var type = ""
+        if buttonRestaurant.tag == 1 {
+            deleteID = modelGetFavouriteByUserResponse?.halalRestuarantResponseData?[index].id ?? "0"
+            type = "rest"
+        }
+        else {
+            deleteID = modelGetFavouriteByUserResponse?.prayerSpacesResponseData?[index].id ?? "0"
+            type = "prayer"
+        }
+        
+        let parameters = [
+            "Id": deleteID,
+            "isMark": false,
+            "type" : type
+        ] as [String : Any]
+       
+        APIs.postAPI(apiName: .postfavouriterestaurants, parameters: parameters, viewController: self) { responseData, success, errorMsg in
+            let model: FindHalalFoodCell.ModelPostFavouriteDeleteResponse? = APIs.decodeDataToObject(data: responseData)
+            self.modelPostFavouriteDeleteResponse = model
+        }
     }
     
     func navigateToAddAddressViewControllerFromEditButton(index: Int) {
@@ -115,18 +145,25 @@ class MyFavouritesViewController: UIViewController {
         self.navigationController?.pushViewController(vc, animated: true)
     }
     func buttonDeleteAddress(index: Int) {
-        let refreshAlert = UIAlertController(title: "User Address", message: "Are you sure you want to delete Address?", preferredStyle: UIAlertController.Style.alert)
-
-        refreshAlert.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { (action: UIAlertAction!) in
-          print("Handle Ok logic here")
-            self.deleteUserAddress(index: index)
-          }))
-
-        refreshAlert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action: UIAlertAction!) in
-          print("Handle Cancel Logic here")
-          }))
-
-        present(refreshAlert, animated: true, completion: nil)
+        showAlertCustomPopup(title: "Delete Favourite!", message: "Are you sure you want to delte favourite item?", iconName: .iconError, buttonNames: [
+            [
+                "buttonName": "Delete",
+                "buttonBackGroundColor": UIColor.white,
+                "buttonTextColor": UIColor.colorRed] as [String : Any],
+            [
+                "buttonName": "Cancel",
+                "buttonBackGroundColor": UIColor.colorRed,
+                "buttonTextColor": UIColor.white]
+        ] as? [[String: AnyObject]]) {buttonName in
+            if buttonName == "Cancel" {
+                
+            }
+            else if buttonName == "Delete" {
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
+                    self.deleteUserAddress(index: index)
+                }
+            }
+        }
     }
     func buttonCheckHandler(index: Int) {
 
@@ -147,14 +184,28 @@ class MyFavouritesViewController: UIViewController {
 extension MyFavouritesViewController: UITableViewDelegate, UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return modelGetFavouriteByUserResponse?.halalRestuarantResponseData?.count ?? 0
+        if buttonRestaurant.tag == 1 {
+            return modelGetFavouriteByUserResponse?.halalRestuarantResponseData?.count ?? 0
+        }
+        else {
+            return modelGetFavouriteByUserResponse?.prayerSpacesResponseData?.count ?? 0
+        }
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "MyFavouriteCell") as! MyFavouriteCell
-        cell.halalRestuarantResponseData = modelGetFavouriteByUserResponse?.halalRestuarantResponseData?[indexPath.row]
+        
+        let recordModel: ModelGetFavouriteByUserResponseData!
+        if buttonRestaurant.tag == 1 {
+            recordModel = modelGetFavouriteByUserResponse?.halalRestuarantResponseData?[indexPath.row]
+        }
+        else {
+            recordModel = modelGetFavouriteByUserResponse?.prayerSpacesResponseData?[indexPath.row]
+        }
+        
+        cell.halalRestuarantResponseData = recordModel
         cell.index = indexPath.row
-        cell.selectedAddressIndex = selectedAddressIndex
+        cell.selectedAddressIndex = selectedIndex
         cell.buttonDeleteHandler = buttonDeleteAddress
         cell.buttonCheckHandler = buttonCheckHandler
         
@@ -163,18 +214,26 @@ extension MyFavouritesViewController: UITableViewDelegate, UITableViewDataSource
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         NSLog ("You selected row: %@ \(indexPath)")
-        selectedAddressIndex = indexPath.row
-        for controller in self.navigationController!.viewControllers as Array {
-            if controller.isKind(of: HomeViewController.self) {
-                if let targetViewController = controller as? HomeViewController {
-                    targetViewController.getuser()
-//                    if let model = modelGetUserAddressResponse?.userAddressesResponseData?[indexPath.row] {
-////                        targetViewController.selectedAddress(modelUserAddressesResponseData: model)
-//                    }
-                    self.navigationController!.popToViewController(controller, animated: true)
-                }
-                break
-            }
+        if buttonRestaurant.tag == 0 {
+            let recordModel = modelGetFavouriteByUserResponse?.halalRestuarantResponseData?[indexPath.row]
         }
+        else {
+            let recordModel = modelGetFavouriteByUserResponse?.prayerSpacesResponseData?[indexPath.row]
+        }
+        
+        
+        selectedIndex = indexPath.row
+//        for controller in self.navigationController!.viewControllers as Array {
+//            if controller.isKind(of: HomeViewController.self) {
+//                if let targetViewController = controller as? HomeViewController {
+//                    targetViewController.getuser()
+////                    if let model = modelGetUserAddressResponse?.userAddressesResponseData?[indexPath.row] {
+//////                        targetViewController.selectedAddress(modelUserAddressesResponseData: model)
+////                    }
+//                    self.navigationController!.popToViewController(controller, animated: true)
+//                }
+//                break
+//            }
+//        }
     }
 }
